+++ src/modules/bhatti/v1/nikasi-gate-pass/nikasi-gate-pass.schema.ts
@@ -43,6 +43,12 @@ export const createNikasiGatePassSchema = z.object({
       .int('Gate pass number must be an integer')
       .positive('Gate pass number must be a positive number'),
 
+    manualGatePassNumber: z.coerce
+      .number()
+      .int('Manual gate pass number must be an integer')
+      .positive('Manual gate pass number must be a positive number')
+      .optional(),
+
     date: z.coerce.date(),
 
     variety: z